wrong data type in entry repository

+++ src/main/java/com/jackmu/slowcapsules/repository/EntryRepository.java
@@ -27,5 +27,5 @@ public interface EntryRepository extends JpaRepository<Entry, Long> {
     List<Entry> findAllBySeriesId(Long seriesId);
     List<Entry> findByEntryId(Long id);
 
-    Entry findBySeriesIdAndOrderNum(Long seriesId, Long orderNum);
+    Entry findBySeriesIdAndOrderNum(Long seriesId, Integer orderNum);
 }
